requirements.txt
llama-cpp-python
fastapi
uvicorn
websockets
python-multipart
requests
bs4